requirements.txt
Django==2.2.12
beautifulsoup4==4.4.1
celery==4.4.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
requests==2.9.1
drf-yasg==1.17.1
